Require  Twitter gem and Dotenv gem
The Dotenv gem allows environment variables to be set in the '.env' file. This is used to set the config info for the Twitter API as environment variables, rather than storing that (sensitive) information in a tracked file.

The Twitter gem is necessary for the primary functionality of the app -- pulling tweets from various users' timelines.