Create new namespace for requiring '/config/environment'
The initial version of the ':generate' namespace used for generating filler tweets required the '/config/environment' file. Because Sinatra has a 'namespace' extension, this requirement interfered with the creation of subsequent namespaces in the Rakefile. (More details in this stackoverflow conversation: http://stackoverflow.com/questions/20638136/undefined-method-desc-for-sinatraapplicationclass)

This commit moves the requiring of '/config/environment' into a separate namespace and task, which can then be called from any subsequent tasks.