Abstract Twitter client config to separate task
Previously, the config was done as part of the rake task to pull the default number of tweets from a user's timeline. This commit moves that functionality to a separate task, which allows for the definition of multiple tasks that rely on a connection to the Twitter API. The :pull_tweets rake task has been modified to depend on the :client_config task. Subsequent tasks can also rely on :client_config to interact with the API.